Правки по скорости.

+++ homework-g594-shevkunov/src/main/java/ru/mipt/java2016/homework/g594/shevkunov/task3/LazyMergedKeyValueStorage.java
@@ -69,7 +69,7 @@ class LazyMergedKeyValueStorage<K, V> implements KeyValueStorage<K, V> {
         }
 
         keeper = new LazyMergedKeyValueStorageKeeper<>(valueSerializator,
-                path + DATA_NAME_PREFIX, DATA_NAME_SUFFIX, header.getDataFilesCount(),
+                path + DATA_NAME_PREFIX, DATA_NAME_SUFFIX, (int) header.getDataFilesCount(),
                 header.createdByConstructor);
     }
