+++ homework-g594-shevkunov/src/main/java/ru/mipt/java2016/homework/g594/shevkunov/task3/LazyMergedKeyValueStorageFileNode.java
@@ -5,19 +5,19 @@ package ru.mipt.java2016.homework.g594.shevkunov.task3;
  * Created by shevkunov on 14.11.16.
  */
 class LazyMergedKeyValueStorageFileNode {
-    private int file;
+    private long file;
     private long offset;
 
-    LazyMergedKeyValueStorageFileNode(int file, long offset) {
+    LazyMergedKeyValueStorageFileNode(long file, long offset) {
         set(file, offset);
     }
 
-    void set(int fileIndex, long offsetValue) {
+    void set(long fileIndex, long offsetValue) {
         this.file = fileIndex;
         this.offset = offsetValue;
     }
 
-    int getFile() {
+    long getFile() {
         return file;
     }
