+++ homework-g594-shevkunov/src/main/java/ru/mipt/java2016/homework/g594/shevkunov/task3/LazyMergedKeyValueStorageKeeper.java
@@ -59,7 +59,7 @@ class LazyMergedKeyValueStorageKeeper<V> {
     }
 
     V read(LazyMergedKeyValueStorageFileNode pointer) throws IOException {
-        return loadFromFile(dataFiles.get(pointer.getFile()), pointer.getOffset());
+        return loadFromFile(dataFiles.get((int) pointer.getFile()), pointer.getOffset());
     }
 
     private long writeToFile(RandomAccessFile out, V value) throws IOException {
